+++ pkg/analyzer/lib/src/dart/element/type_system.dart
@@ -273,7 +273,7 @@ abstract class TypeSystem implements public.TypeSystem {
     if (type is! InterfaceType) return null;
     if (genericType is! InterfaceType) return null;
 
-    var asInstanceOf = (type as InterfaceTypeImpl)
+    var asInstanceOf = (type as InterfaceType)
         .asInstanceOf((genericType as InterfaceType).element);
 
     if (asInstanceOf != null) {